create table work under progress

+++ code/db_utils/create-utils.go
@@ -1,6 +1,7 @@
 package dbutils
 
 import (
+	"errors"
 	"fmt"
 	"os"
 	"path/filepath"
@@ -23,6 +24,13 @@ func CreateDB(dbName string) (err error) {
 		loghelper.LogError(err.Error())
 		return
 	}
+
+	db := Database{
+		Name: dbName,
+	}
+
+	Databases[dbName] = db
+
 	loghelper.LogInfo(fmt.Sprintf("Database %v created", dbName))
 	return
 }
@@ -37,11 +45,25 @@ func DropDB(dbName string) (err error) {
 	pathToDBEngineFiles := filepath.Join(home, "VectorDB")
 	newDB := filepath.Join(pathToDBEngineFiles, dbName)
 
+	var confirmation string
+
+	fmt.Printf("Are you sure to delete the database %v (y/n): ", dbName)
+	fmt.Scan(&confirmation)
+
+	if confirmation == "n\n" || confirmation == "N\n" {
+		errMsg := "Aborting operation...."
+		loghelper.LogError(errMsg)
+		err = errors.New(errMsg)
+		return
+	}
+
 	if err = os.RemoveAll(newDB); err != nil {
 		loghelper.LogError(err.Error())
 		return
 	}
+  
+  // Refresh database 
+
 	loghelper.LogInfo(fmt.Sprintf("Database %v deleted", dbName))
 	return
 }
-
